LayoutTests/platform/gtk/accessibility/canvas-description-and-role-expected.txt

This test makes sure that a canvas with and without fallback content each has the right role and description.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS axContainer.childrenCount is 2
Canvas 1 title: AXTitle: Canvas label
Canvas 1 role: AXRole: AXCanvas
Canvas 2 title: AXTitle: 
Canvas 2 role: AXRole: AXCanvas
PASS successfullyParsed is true

TEST COMPLETE
